알람 기능 수정

+++ src/main/java/com/tickettimer/backendserver/controller/AlarmController.java
@@ -6,6 +6,8 @@ import com.tickettimer.backendserver.domain.musical.Actor;
 import com.tickettimer.backendserver.domain.musical.Musical;
 import com.tickettimer.backendserver.domain.musical.MusicalNotice;
 import com.tickettimer.backendserver.dto.ActorRequestDto;
+import com.tickettimer.backendserver.dto.AlarmRequest;
+import com.tickettimer.backendserver.dto.AlarmResponse;
 import com.tickettimer.backendserver.dto.ResultResponse;
 import com.tickettimer.backendserver.service.AlarmService;
 import com.tickettimer.backendserver.service.JwtService;
@@ -20,6 +22,9 @@ import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+import java.util.stream.Collectors;
+
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/alarms")
@@ -28,18 +33,35 @@ public class AlarmController {
     private final MusicalNoticeService musicalNoticeService;
     private final MemberService memberService;
     private final JwtService jwtService;
+    @GetMapping
+    public ResponseEntity<ResultResponse> postAlarm(
+            HttpServletRequest request
+    ) {
+        Long id = (Long) request.getAttribute("id");
+        Member member = memberService.findById(id);
+        List<Alarm> alarms = member.getAlarms();
+        List<AlarmResponse> response = alarms.stream().map(
+                m -> AlarmResponse.builder().id(m.getId()).musicalNotice(m.getMusicalNotice()).alarmTimes(m.getAlarmTimes()).build()
+        ).collect(Collectors.toList());
+
+        ResultResponse res = ResultResponse.builder()
+                .code(HttpStatus.CREATED.value())
+                .message("등록한 알람을 가져왔습니다.")
+                .result(response).build();
+        return new ResponseEntity<>(res, HttpStatusCode.valueOf(res.getCode()));
+    }
     @PostMapping
     public ResponseEntity<ResultResponse> postAlarm(
             HttpServletRequest request,
-            @RequestParam("musicalNoticeId") String musicalNoticeId
+            @RequestBody AlarmRequest alarmRequest
     ) {
         Long id = (Long) request.getAttribute("id");
-        System.out.println("id = " + id);
-        MusicalNotice musicalNotice = musicalNoticeService.findById(musicalNoticeId);
+        MusicalNotice musicalNotice = musicalNoticeService.findById(alarmRequest.getMusicalNoticeId());
         Member member = memberService.findById(id);
         Alarm alarm = Alarm.builder()
                 .musicalNotice(musicalNotice)
                 .member(member)
+                .alarmTimes(alarmRequest.getAlarmTimes())
                 .build();
 
         // Alarm 저장
@@ -51,16 +73,31 @@ public class AlarmController {
                 .result(musicalNotice).build();
         return new ResponseEntity<>(res, HttpStatusCode.valueOf(res.getCode()));
     }
-    @DeleteMapping
+    @DeleteMapping("/{id}")
     public ResponseEntity<ResultResponse> deleteAlarm(
             HttpServletRequest request,
-            @RequestParam("id") Long alarmId
+            @PathVariable("id") Long alarmId
     ) {
+        // 알람을 등록한 사람이 요청한 것인지 확인
+        Long id = (Long) request.getAttribute("id");
+        Member member = memberService.findById(id);
+        Alarm alarm = alarmService.findById(alarmId);
+
+        // 다른 사람이 요청했다면 거부
+        if (member != alarm.getMember()) {
+            ResultResponse res = ResultResponse.builder()
+                    .code(HttpStatus.UNAUTHORIZED.value())
+                    .message(alarmId+ " : 알람을 삭제할 권한이 없습니다.")
+                    .build();
+            return new ResponseEntity<>(res, HttpStatusCode.valueOf(res.getCode()));
+
+        }
         alarmService.delete(alarmId);
         // 결과 반환
         ResultResponse res = ResultResponse.builder()
-                .code(HttpStatus.CREATED.value())
-                .message(alarmId+ " : 뮤지컬 정보를 삭제했습니다.")
+                .code(HttpStatus.OK.value())
+                .message(alarmId+ " : 알람을 삭제했습니다.")
+                .result(alarm.getMusicalNotice())
                 .build();
         return new ResponseEntity<>(res, HttpStatusCode.valueOf(res.getCode()));
     }
